commit 8e9352593f28b1a84101b11e1c6db6aabf187a1c
    5880 Increase IOV_MAX to at least 1024
commit f3e1a610966249cf3c18b6f431eef3fac2ed90e4
    12279 ::arc_compression_stats generates errors
commit 3c6ffbab91273559b511d95f850d7b2d9cd2a3c5
    11958 need topo maps for the SMCI,SYS-2028U-E1CNRT+
    11959 extend disk topo plugin to enumerate nvme devices

	Conflicts:
	usr/src/uts/common/syscall/rw.c
	usr/src/uts/common/sys/socket.h

+++ usr/src/uts/common/syscall/rw.c
@@ -1087,7 +1087,7 @@ preadv(int fdes, struct iovec *iovp, int iovcnt, off_t offset,
 		}
 	}
 
-	if ((bcount = count) < 0) {
+	if ((bcount = (ssize_t)count) < 0) {
 		if (aiovlen != 0)
 			kmem_free(aiov, aiovlen);
 		return (set_errno(EINVAL));
@@ -1288,7 +1288,7 @@ pwritev(int fdes, struct iovec *iovp, int iovcnt, off_t offset,
 		}
 	}
 
-	if ((bcount = count) < 0) {
+	if ((bcount = (ssize_t)count) < 0) {
 		if (aiovlen != 0)
 			kmem_free(aiov, aiovlen);
 		return (set_errno(EINVAL));
